+++ src/components/ComponentWrapper/ComponentWrapper.tsx
@@ -54,14 +54,6 @@ export const ComponentWrapper: React.FC<React.PropsWithChildren<any>> = chakra(
         },
       });
 
-      // Note: Typescript warns when there are no children passed and
-      // doesn't compile. This is meant to log in non-Typescript apps.
-      if (!hasChildren) {
-        console.warn(
-          "NYPL Reservoir ComponentWrapper: No children were passed."
-        );
-      }
-
       return (
         <Box id={`${id}-wrapper`} ref={ref} __css={styles} {...rest}>
           {finalHeadingText}