+++ CHANGELOG.md
@@ -52,6 +52,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Removes the use of the `window`'s `scrollIntoView` dependency in the `Tabs` component.
 - Removes the `tableTextSize` prop from the `Table` component in favor of using the native `fontSize` CSS prop.
 - Removes `Children.map` logic from `Button`, `ButtonGroup`, `Card`, `CheckboxGroup`, `List`, `RadioGroup`, and `Tooltip` to reduce complexity and improve performance.
+- Removes warnings validating presence or type of children from `SubNav`, `ComponentWrapper`, `Form`, `Heading`, `Icon`, `Logo`, and `Text`.
+- Removes the `FilterBar` component and `useFilterBar` hook.
 - Removes `noSpace` prop from Heading component.
 - Removes `noSpace` prop from Text component.
 - Removes default spacing values set on the `DatePicker`, `Heading`, `HelperErrorText`, `HorizontalRule`, `Label`, `List`, `StyledList`, and `Text` components.